# openapi .json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Define an API",
    "title": "Restful API(connect two databases)",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "url": "http://www.example.com/support",
      "email": "a167829435@gmail.com"
    }
  },
  "paths": {
...
    "/v1/mssql/getall": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "從events取得所有資料",
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Event"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "get all data from events",
        "tags": [
          "Event"
        ]
      }
    },
    "/v1/mssql/getsome/{key1}/{key2}/{key3}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "從events取得部分資料",
        "parameters": [
          {
            "name": "key1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Event"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "The user does not exist!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "get some data from events",
        "tags": [
          "Event"
        ]
      }
    },
    "/v1/mssql/addvalue": {
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "加入數值至events",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/models.Event"
              }
            }
          },
          "description": "add data",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Event"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "add value to events",
        "tags": [
          "Event"
        ]
      }
    },
    "/v1/mssql/update": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "更新events資料",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/model.mssqlupdate"
              },
              "example": {
                "key1": "12345",
                "birth": 83
              }
            }
          },
          "description": "update data",
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Successfully update"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "update value",
        "tags": [
          "Event"
        ]
      }
    },
    "/v1/mssql/delete/{key1}/{key2}/{key3}": {
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "刪除events資料",
        "parameters": [
          {
            "name": "key1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                },
                "example": "Successfully delete"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "delete value",
        "tags": [
          "Event"
        ]
      }
    },
    "/v1/join/getall": {
      "get": {
        "description": "取得所有資料",
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.JoinTable"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2,
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "取得關聯表部分資料",
        "tags": [
          "JoinTable"
        ]
      }
    },
    "/v1/join/getsome/{key1}/{key2}/{key3}": {
      "get": {
        "parameters": [
          {
            "name": "key1",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key2",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "key3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "description": "取得部分資料",
        "responses": {
          "200": {
            "description": "Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.JoinTable"
                },
                "example": {
                  "Key1": "10600",
                  "Key2": "10604240020017",
                  "Key3": "A24WZ",
                  "Number": 100,
                  "Gender": 1,
                  "Birth": 83,
                  "Injury_degree": "3",
                  "Injury_position": 10,
                  "Protection": 1,
                  "Phone": 2,
                  "Person": "B03",
                  "Car": "7",
                  "Action_status": 1,
                  "Qualification": 2,
                  "License": 3,
                  "Drinking": 1,
                  "Hit": 2,
                  "City": "台中",
                  "Position": "國道一號",
                  "Lane": "外側",
                  "Death": "0",
                  "Injured": "2",
                  "Death_exceed": "0",
                  "Weather": "晴",
                  "Light": "日間光線",
                  "Time_year": 2017,
                  "Time_month": "4",
                  "Accident_chinese": "車與車",
                  "Anecdote_chinese": "駕駛人"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "The user does not exist!"
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/models.Error"
                },
                "example": "Serve(database) error!"
              }
            }
          }
        },
        "summary": "取得關聯表部分資料",
        "tags": [
          "JoinTable"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1"
    }
  ],
  "components": {
...
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "model.mssqlupdate": {
        "properties": {
          "death": {
            "type": "string"
          },
          "key1": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.Error": {
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "models.Event": {
        "properties": {
          "accident_chinese": {
            "type": "string"
          },
          "anecdote_chinese": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "death": {
            "type": "string"
          },
          "death_exceed": {
            "type": "string"
          },
          "injured": {
            "type": "string"
          },
          "key1": {
            "type": "string"
          },
          "key2": {
            "type": "string"
          },
          "key3": {
            "type": "string"
          },
          "lane": {
            "type": "string"
          },
          "light": {
            "type": "string"
          },
          "position": {
            "type": "string"
          },
          "time_month": {
            "type": "string"
          },
          "time_year": {
            "type": "integer"
          },
          "weather": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "models.JoinTable": {
        "properties": {
          "accident_chinese": {
            "type": "string"
          },
          "action_status": {
            "type": "integer"
          },
          "anecdote_chinese": {
            "type": "string"
          },
          "birth": {
            "type": "integer"
          },
          "car": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "death": {
            "type": "string"
          },
          "death_exceed": {
            "type": "string"
          },
          "drinking": {
            "type": "integer"
          },
          "gender": {
            "type": "integer"
          },
          "hit": {
            "type": "integer"
          },
          "injured": {
            "type": "string"
          },
          "injury_degree": {
            "type": "string"
          },
          "injury_position": {
            "type": "integer"
          },
          "key1": {
            "type": "string"
          },
          "key2": {
            "type": "string"
          },
          "key3": {
            "type": "string"
          },
          "lane": {
            "type": "string"
          },
          "license": {
            "type": "integer"
          },
          "light": {
            "type": "string"
          },
          "number": {
            "type": "integer"
          },
          "person": {
            "type": "string"
          },
          "phone": {
            "type": "integer"
          },
          "position": {
            "type": "string"
          },
          "protection": {
            "type": "integer"
          },
          "qualification": {
            "type": "integer"
          },
          "time_month": {
            "type": "string"
          },
          "time_year": {
            "type": "integer"
          },
          "weather": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  }
}